Refactored into single API call and added test
